backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
SQLAlchemy==2.0.30
python-dotenv==1.0.1
psycopg2-binary==2.9.9
openai==1.30.0
httpx==0.27.0
python-multipart==0.0.9
pydantic-settings==2.2.1
reportlab==4.1.0
azure-identity==1.16.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
aiofiles
pydantic[email]
azure-keyvault-secrets
msgraph-sdk
ffmpeg-python
aiohttp
websockets
Pillow
numpy
celery[gevent]==5.3.6
redis[hiredis]==4.6.0
kombu==5.3.4
flower==2.0.1
gevent==23.9.1
boto3
asyncpg
psutil>=5.9.0
gunicorn==21.2.0
pytz==2024.1
alembic==1.13.1 